+++ NEWS.md
@@ -6,6 +6,7 @@
 
 ## Bug fix
 - For CTIS, `dbFindIdsUniqueTrials()` returns now only the trial record with the highest resubmission number
+- For EUCTR, correct edge case downloading third-country protocol-related information 
 - Fixed links between help pages and vignettes
 
 # ctrdata 1.25.1